Add support to specify UID and GID for rtorrent user
The rtorrent user will now be created with UID=1000 and GID=100
if the user doesn't set the following environment variables.
USR_ID, GRP_ID can be used to specify what UID and GID the rtorrent
user will use in the container.
This feature makes it easier to deploy on host systems where the
UID and GID in the container doesn't exist on the host.