Make it possible to use external rtorrent.rc setting file
Instead of the settings included in the image, it's now possible to
use an external settings file put in the /downloads/.rtorrent folder.
This makes it easier to configure rtorrent the way you'd like.

The file is symlinked inside the container so if it doesn't exist
the default settings will be used.